requirements.txt
discord.py>=2.3.2
cloudscraper>=1.2.71
beautifulsoup4>=4.12.2
aiosqlite>=0.19.0
python-dotenv>=1.0.0

